+++ zapit/+zapit/+gui/+stimConfigEditor/@controller/returnStimConfigStructure.m
@@ -20,7 +20,7 @@ function stimC = returnStimConfigStructure(obj)
     %  https://github.com/Zapit-Optostim/zapit/issues/9
 
     pointAttributes.laserPowerInMW = obj.LaserPowermWSpinner.Value;
-    pointAttributes.stimDutyCycleHz = obj.StimFreqHzSpinner.Value;
+    pointAttributes.stimModulationFreqHz = obj.StimFreqHzSpinner.Value;
     pointAttributes.offRampDownDuration_ms = obj.RampdownmsSpinner.Value;
 
     for ii=1:length(obj.pAddedPoints)